# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: MebloPlan Scanner API
  version: 0.1.0
paths:
  /health:
    get:
      summary: Service health check.
      responses:
        '200':
          description: Service is up.
  /api/scans:
    post:
      summary: Upload scan data and convert to GLB.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: OBJ, PLY, USD/USDA/USDZ
                meta:
                  type: string
                  description: Additional metadata encoded as JSON or URL-encoded string.
      responses:
        '200':
          description: Conversion completed.
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - url
                properties:
                  id:
                    type: string
                    format: uuid
                    description: Unique scan ID.
                  url:
                    type: string
                    format: uri
                    description: Full URL to download the GLB file.
                    example: http://localhost:4000/api/scans/123e4567-e89b-12d3-a456-426614174000/room.glb
        '400':
          description: Bad request.
        '401':
          description: Unauthorized.
        '429':
          description: Too many requests. Conversion queue is full.
        '500':
          description: Server error.
  /api/scans/{id}/info:
    get:
      summary: Retrieve scan metadata by scan id.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: Unique scan ID.
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Metadata saved during upload.
          headers:
            Cache-Control:
              description: Cache policy for the metadata.
              schema:
                type: string
              example: public, max-age=86400, immutable
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Bad request.
        '401':
          description: Unauthorized.
        '404':
          description: Not found.
        '429':
          description: Too many requests.
        '500':
          description: Server error.
  /api/scans/{id}/room.glb:
    get:
      summary: Download converted GLB model by scan id.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: Unique scan ID.
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: GLB file.
          headers:
            Cache-Control:
              description: Cache policy for the model.
              schema:
                type: string
              example: public, max-age=86400, immutable
          content:
            model/gltf-binary:
              schema:
                type: string
                format: binary
        '400':
          description: Bad request.
        '401':
          description: Unauthorized.
        '404':
          description: Not found.
        '429':
          description: Too many requests.
        '500':
          description: Server error.
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer